requirements.txt
langchain_community
langchain_core
langchain-openai
langchain
python-dotenv
streamlit
